+++ spython/utils/terminal.py
@@ -134,13 +134,16 @@ def stream_command(cmd, no_newline_regexp="Progess", sudo=False, sudo_options=No
     """
     cmd = _process_sudo_cmd(cmd, sudo, sudo_options)
 
-    process = subprocess.Popen(cmd, stdout=subprocess.PIPE, universal_newlines=True)
+    process = subprocess.Popen(
+        cmd, stdout=subprocess.PIPE, stderr=subprocess.PIPE, universal_newlines=True
+    )
     for line in iter(process.stdout.readline, ""):
         if not re.search(no_newline_regexp, line):
             yield line
     process.stdout.close()
     return_code = process.wait()
     if return_code:
+        print(process.stderr.read(), file=sys.stderr)
         raise subprocess.CalledProcessError(return_code, cmd)